fastapi/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
google-generativeai==0.3.2
python-dotenv==1.0.1
manim==0.17.3
pydub==0.25.1
ffmpeg-python==0.2.0